// js/app/src/pages/playground/evaluators/__generated__/EvaluatorSlotEditorSaveCodeMutation.graphql.ts
export type CreateCodeEvaluatorInput = {
  description?: string | null;
  inputMapping?: EvaluatorInputMappingInput | null;
  language: Language;
  name: string;
  outputConfigs?: ReadonlyArray<AnnotationConfigInput> | null;
  sandboxConfigId: string;
  sourceCode: string;
};
// ...
export type EvaluatorSlotEditorSaveCodeMutation$variables = {
  input: CreateCodeEvaluatorInput;
};
export type EvaluatorSlotEditorSaveCodeMutation$data = {
  readonly createCodeEvaluator: {
    readonly evaluator: {
      readonly id: string;
    };
  };
};
export type EvaluatorSlotEditorSaveCodeMutation = {
  response: EvaluatorSlotEditorSaveCodeMutation$data;
  variables: EvaluatorSlotEditorSaveCodeMutation$variables;
};